Adding endpoint to get project summary from Open AI API

+++ src/main/java/com/goldengit/restclient/schema/WeekOfCommit.java
@@ -2,5 +2,5 @@ package com.goldengit.restclient.schema;
 
 public class WeekOfCommit {
     public long week;
-    public int total;
+    public long total;
 }
